Feat: 식당, 리뷰, 사용자 정보 출력 기능 개발

+++ web-server/src/main/java/com/example/webserver/repository/JdbcReviewRepository.java
@@ -41,7 +41,7 @@ public class JdbcReviewRepository {
                         ps.setInt(parameterIndex, review.getReviewScore());
                         parameterIndex = parameterIndex + 1;
                         //set normReviewScore
-                        ps.setInt(parameterIndex, review.getNormReviewScore());
+                        ps.setDouble(parameterIndex, review.getNormReviewScore());
                         parameterIndex = parameterIndex + 1;
                     }
